fix test

+++ x/concentrated-liquidity/position_test.go
@@ -624,7 +624,7 @@ func (s *KeeperTestSuite) TestFungifyChargedPositions() {
 			if test.expectedErr != nil {
 				s.Require().Error(err)
 				s.Require().ErrorIs(err, test.expectedErr)
-				s.Require().Equal(0, newPositionId)
+				s.Require().Equal(uint64(0), newPositionId)
 			} else {
 				s.Require().NoError(err)
 				s.Require().Equal(test.expectedNewPositionId, newPositionId)
